Complete CRUD and authentication and authorization, thanks Brad!

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc   get goals
 // @route  GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req,res) => { //get
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -17,7 +18,8 @@ const setGoal =  asyncHandler(async (req,res) => { //set
         throw new Error('No text field, add one please!')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -27,9 +29,22 @@ const setGoal =  asyncHandler(async (req,res) => { //set
 // @access Private
 const updateGoal =  asyncHandler(async (req,res) => { //set
     const goal = await Goal.findById(req.params.id)
+    
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found!')
+    }
+    
+    const user = await User.findById(req.user.id)
+    //check user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure user is allowed to update
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -45,6 +60,17 @@ const deleteGoal =  asyncHandler(async (req,res) => { //set
     if(!goal) {
         res.status(400)
         throw new Error('Goal not found!')
+    }
+    const user = await User.findById(req.user.id)
+    //check user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    //make sure user is allowed to update
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await Goal.deleteOne()
     res.status(200).json({id: req.params.id})
